librobinhood: fix gcc warnings

+++ librobinhood/src/backends/posix/posix.c
@@ -106,7 +106,7 @@ retry:
         errno = save_errno;
         return NULL;
     }
-    if (rc == size) {
+    if ((size_t)rc == size) {
         /* Output may have been truncated, try a bigger size to check */
         free(symlink);
 
@@ -148,7 +148,7 @@ ceil2(size_t number)
         return number;
 
     shift = sizeof(number) * 8 - __builtin_clzll(number);
-    return shift >= sizeof(number) * 8 ? number : 1 << shift;
+    return shift >= sizeof(number) * 8 ? number : 1u << shift;
 }
 
 static ssize_t
@@ -181,7 +181,7 @@ retry:
                 }
             }
 
-            if (length <= buflen)
+            if ((size_t)length <= buflen)
                 /* The list of xattrs must have shrunk in between both calls */
                 goto retry;
 
@@ -238,7 +238,7 @@ getxattrs(char *proc_fd_path, struct rbh_value_pair **_pairs,
         return -1;
 
     name = names;
-    for (size_t i = 0; i < count; i++, name += strlen(name) + 1) {
+    for (size_t i = 0; i < (size_t)count; i++, name += strlen(name) + 1) {
         struct rbh_value_pair *pair = &pairs[i - skipped];
         char buffer[XATTR_VALUE_MAX_VFS_SIZE];
         ssize_t length;
@@ -272,7 +272,7 @@ getxattrs(char *proc_fd_path, struct rbh_value_pair **_pairs,
                 return -1;
             }
         }
-        assert(length <= sizeof(buffer));
+        assert((size_t)length <= sizeof(buffer));
 
         pair->value = create_value_from_xattr(name, buffer, length, xattrs);
         if (pair->value == NULL)
